Conflicts:
	src/Sylius/Bundle/AddressingBundle/Resources/config/validation.xml
	src/Sylius/Bundle/UserBundle/Controller/UserController.php

+++ src/Sylius/Component/Core/Uploader/ImageUploader.php
@@ -9,13 +9,18 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace Sylius\Component\Core\Uploader;
 
 use Gaufrette\Filesystem;
 use Sylius\Component\Core\Model\ImageInterface;
+use Symfony\Component\HttpFoundation\File\File;
+use Webmozart\Assert\Assert;
 
 class ImageUploader implements ImageUploaderInterface
 {
+    /** @var Filesystem */
     protected $filesystem;
 
     public function __construct(Filesystem $filesystem)
@@ -23,20 +28,28 @@ class ImageUploader implements ImageUploaderInterface
         $this->filesystem = $filesystem;
     }
 
-    public function upload(ImageInterface $image)
+    /**
+     * {@inheritdoc}
+     */
+    public function upload(ImageInterface $image): void
     {
         if (!$image->hasFile()) {
             return;
         }
 
-        if (null !== $image->getPath()) {
+        $file = $image->getFile();
+
+        /** @var File $file */
+        Assert::isInstanceOf($file, File::class);
+
+        if (null !== $image->getPath() && $this->has($image->getPath())) {
             $this->remove($image->getPath());
         }
 
         do {
-            $hash = md5(uniqid(mt_rand(), true));
-            $path = $this->expandPath($hash.'.'.$image->getFile()->guessExtension());
-        } while ($this->filesystem->has($path));
+            $hash = bin2hex(random_bytes(16));
+            $path = $this->expandPath($hash . '.' . $file->guessExtension());
+        } while ($this->isAdBlockingProne($path) || $this->filesystem->has($path));
 
         $image->setPath($path);
 
@@ -46,12 +59,19 @@ class ImageUploader implements ImageUploaderInterface
         );
     }
 
-    public function remove($path)
+    /**
+     * {@inheritdoc}
+     */
+    public function remove(string $path): bool
     {
-        return $this->filesystem->delete($path);
+        if ($this->filesystem->has($path)) {
+            return $this->filesystem->delete($path);
+        }
+
+        return false;
     }
 
-    private function expandPath($path)
+    private function expandPath(string $path): string
     {
         return sprintf(
             '%s/%s/%s',
@@ -59,5 +79,18 @@ class ImageUploader implements ImageUploaderInterface
             substr($path, 2, 2),
             substr($path, 4)
         );
+    }
+
+    private function has(string $path): bool
+    {
+        return $this->filesystem->has($path);
+    }
+
+    /**
+     * Will return true if the path is prone to be blocked by ad blockers
+     */
+    private function isAdBlockingProne(string $path): bool
+    {
+        return strpos($path, 'ad') !== false;
     }
 }
